Connect the Application with store and Seperate the file with folder structure
