+++ src/actions/actionTypes.js
@@ -1,0 +1,3 @@
+export const SET_INTERVAL = "SET INTERVAL";
+export const SET_OPTIONLIST = "SET OPTIONLIST";
+// export const SET_LOOP = "SET LOOP";
