+++ src/store.js
@@ -1,0 +1,4 @@
+import {createStore} from "redux";
+import screenLoopReducer from "./reducers/screenLoopReducer";
+
+export default createStore(screenLoopReducer);
